Added useLoginToast hook
- fixed update dependencies action and added worflow_dispatch
- added hooks and types folders
- moved files around to the new folders
- added useLoginToast hook/toast
