+++ .github/workflows/update-dependencies.yaml
@@ -1,6 +1,7 @@
 name: Dependency Update Check
 
 on:
+  workflow_dispatch:
   # Schedule the job to run on the first day of every month at midnight
   schedule:
     - cron: '0 0 1 * *'
@@ -38,7 +39,7 @@ jobs:
         run: npx npm-check-updates --upgrade --jsonUpgraded
         id: check
 
-      - name: Create a draft pull request
+      - name: Create a pull request
         uses: peter-evans/create-pull-request@v5
         with:
           token: ${{ secrets.GH_PAGES_TOKEN }}
@@ -51,4 +52,3 @@ jobs:
             ```
           labels: dependencies
           base: dev
-          draft: true
