+++ src/components/primitives/Toaster.tsx
@@ -1,4 +1,4 @@
-import { useToast } from '~/lib/use-toast';
+import { useToast } from '~/hooks/use-toast';
 import {
   Toast,
   ToastClose,
