+++ src/pages/_app.tsx
@@ -9,7 +9,7 @@ import Spinner from '~/components/spinners/Spinner';
 import { ThemeProvider } from '~/components/utility/ThemeProvider';
 import { api } from '~/lib/api';
 import printConsoleLogo from '~/lib/printConsoleLogo';
-import { usePageLoading } from '~/lib/use-page-loading';
+import { usePageLoading } from '~/hooks/use-page-loading';
 import '~/styles/globals.css';
 
 const inter = Inter({
