+++ src/pages/organization/[id].tsx
@@ -2,7 +2,7 @@ import Image from 'next/image';
 import { useSearchParams } from 'next/navigation';
 import { type GetServerSideProps, type NextPage } from 'next/types';
 import PageLayout from '~/components/layouts/PageLayout';
-import { type IOrganizationData } from '~/lib/petfinderTypes';
+import { type IOrganizationData } from '~/types/petfinderTypes';
 import { type PetfinderOauth } from '../results';
 
 type IOrganizationProfilePage = {
