+++ src/pages/search.tsx
@@ -6,7 +6,7 @@ import PageLayout from '~/components/layouts/PageLayout';
 import BackgroundWave from '~/components/utility/BackgroundWave';
 import SearchResults from '~/components/utility/SearchResults';
 import { TypeOfResults } from '~/lib/constants';
-import type IAnimalData from '~/lib/petfinderTypes';
+import type IAnimalData from '~/types/petfinderTypes';
 
 export interface IResults {
   animals?: IAnimalData[];
