+++ src/types/next-auth.d.ts
@@ -1,0 +1,21 @@
+import { type DefaultSession } from 'next-auth';
+
+/**
+ * Module augmentation for `next-auth` types. Allows us to add custom properties to the `session`
+ * object and keep type safety.
+ *
+ * @see https://next-auth.js.org/getting-started/typescript#module-augmentation
+ */
+declare module 'next-auth' {
+  interface Session extends DefaultSession {
+    user: {
+      id: string;
+      role?: string;
+    } & DefaultSession['user'];
+  }
+
+  interface User {
+    id: string;
+    role?: string;
+  }
+}
